feat: added normalization
